pyproject.toml
[project]
name = "app_version_inventory"
version = "1.0.0-dev"
dependencies = [
    "click",
    "kubernetes",
    "pyyaml"
]
requires-python = ">=3.12"

[project.optional-dependencies]
dev = [
     "ruff"
]
test = [
    "pytest",
    "pytest-cov"
]

[project.scripts]
app-version-inventory = "app_version_inventory.cli:main"